// frontend/src/app/shared/proxy/models/SaveSetupConfigurationRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SaveSetupConfigurationRequest = {
  databaseConnectionString?: string | null;
  redisConnectionString?: string | null;
  keycloakAuthority?: string | null;
  keycloakRealm?: string | null;
  keycloakClientId?: string | null;
  keycloakClientSecret?: string | null;
};